Base de datos JDBC

+++ 09_agenda_contactos/src/service/ContactoService.java
@@ -3,8 +3,11 @@ package service;
 import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.PreparedStatement;
+import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
+import java.util.ArrayList;
+import java.util.List;
 
 import model.Contacto;
 
@@ -24,9 +27,13 @@ public class ContactoService {
 	String usuario = "root";
 	String password = "root";
 	
-	public void agregarContacto(Contacto contacto) {
-		
-		try (Connection con = DriverManager.getConnection(cadenaConexion, usuario, password);) {
+	// no se admiten contactos con email duplicado
+	public boolean nuevoContacto(Contacto contacto) {
+		if(existeContactoPorEmail(contacto.getEmail()) !=null) {
+			System.out.println("ya existe ese email");
+			return false;
+		}
+		try (Connection con = DriverManager.getConnection(cadenaConexion, usuario, password);) {	
 			String sql = "insert into contactos(nombre, email, edad) values(?,?,?)";	
 			PreparedStatement ps = con.prepareStatement(sql);
 			ps.setString(1,  contacto.getNombre());
@@ -34,24 +41,32 @@ public class ContactoService {
 			ps.setInt(3,  contacto.getEdad());
 			ps.execute(); 
 			System.out.println("¡Registro añadido!");
+			return true;
+			
 		}catch (SQLException ex) {
-			ex.printStackTrace();
-		}
-	}
-	
-	public boolean eliminarContacto (String email) {
-		// delete from empleados where codigodep="BB45"
-		try (Connection con = DriverManager.getConnection(cadenaConexion, usuario, password);) {
-			String sql = "delete from contactos where email=?";	
-			PreparedStatement ps = con.prepareStatement(sql);
-			ps.setString(1, email);
-			return ps.executeUpdate()>0;
-		//	Statement st = con.createStatement();			
-			//int resultado = st.executeUpdate(sql);
-		}catch (SQLException ex) {
-			ex.printStackTrace();
+			ex.printStackTrace();			
 			return false;
 		}
+		
+	}
+	
+	// devuelve el contacto eliminado. Si no se ha eliminado ninguno, devuelve null
+	public Contacto eliminarContacto (String email) {
+		Contacto contacto = existeContactoPorEmail(email);
+		if (contacto==null) {
+			return null;
+		}
+		try (Connection con = DriverManager.getConnection(cadenaConexion, usuario, password);) {
+			String sql = "delete from contactos where email=?";
+			PreparedStatement ps = con.prepareStatement(sql);
+			ps.setString(1, email);
+			System.out.println("contacto borrado");
+			return contacto;
+			
+		}catch (SQLException ex) {
+			ex.printStackTrace();
+			return null;
+		}		
 	}
 	
 	
@@ -68,5 +83,69 @@ public class ContactoService {
 			return false;
 		}
 	}
+	
+	// buscar contacto por su clave primaria 
+	public Contacto buscarContactoPorId(int idContacto) {
+		try (Connection con = DriverManager.getConnection(cadenaConexion, usuario, password);) {			
+			String sql = "select * from contactos where idContacto=?";
+			PreparedStatement st = con.prepareStatement(sql);
+			st.setInt(1, idContacto);			
+			ResultSet rs = st.executeQuery(); // no se le pasa sql porque se le pasa ya al prepareStatement
+			
+			// debemos movernos a la primera, y única fila, para poder extraer
+			if(rs.next()) {
+				return new  Contacto(rs.getInt("idContacto"), rs.getString("nombre"), rs.getString("email"), rs.getInt("edad"));
+			}
+			// otra forma si no hago el select con where
+//			while(rs.next()) {
+//				if((rs.getInt("idContacto")) == (idContacto)) {
+//					return new Contacto(rs.getInt("idContacto"), rs.getString("nombre"), rs.getString("email"), rs.getInt("edad"));
+//				}
+//			}			
+			
+		}catch (SQLException ex) {
+			ex.printStackTrace();
+		}
+		return null;
+	}
+	
+	// recupera todos los contactos
+	public List<Contacto> getContactos() {
+		List<Contacto> contactos = new ArrayList<Contacto>();
+		try (Connection con = DriverManager.getConnection(cadenaConexion, usuario, password);) {			
+			String sql = "select * from contactos";
+			Statement st = con.createStatement();
+			ResultSet rs = st.executeQuery(sql);
+			while(rs.next()) {
+				contactos.add(new Contacto(rs.getInt("idContacto"), rs.getString("nombre"),
+						rs.getString("email"), rs.getInt("edad")));
+			}			
+		}catch (SQLException ex) {
+			ex.printStackTrace();
+		}
+		return contactos;
+	}
+	
+	private Contacto existeContactoPorEmail(String email) {
+		try(Connection con=DriverManager.getConnection(cadenaConexion,usuario,password)){		
+			String sql="select * from contactos where email=?";
+			PreparedStatement st=con.prepareStatement(sql);
+			st.setString(1, email);
+			ResultSet rs=st.executeQuery();
+			//debemos movernos a la primera y única fila, para poder extraer
+			//el valor de dicha fila
+			if(rs.next()) {
+				return new Contacto(rs.getInt("idContacto"),
+							rs.getString("nombre"),
+							rs.getString("email"),
+							rs.getInt("edad"));
+			}
+			return null;
+		}catch(SQLException ex) {
+			ex.printStackTrace();
+			return null;
+		}
+		
+	} 
 	
 }
